+++ src/pages/apply/Apply.tsx
@@ -2,8 +2,8 @@ import React from 'react';
 import { Helmet } from 'react-helmet';
 import { Container, Box } from '@material-ui/core';
 import ApplyGrid from './ApplyGrid';
-import Hero from './Hero';
-import { ReactComponent as SVG } from '../svg/team.svg';
+import Hero from '../../components/Hero';
+import { ReactComponent as SVG } from '../../svg/team.svg';
 
 const Apply: React.FC = () => {
   React.useEffect(() => {
