+++ src/pages/people/Member.tsx
@@ -1,0 +1,114 @@
+import React from 'react';
+import { makeStyles } from '@material-ui/core/styles';
+import { Paper, Grid, Typography } from '@material-ui/core';
+// @ts-ignore
+import { LazyLoadImage } from 'react-lazy-load-image-component';
+
+const useStyles = makeStyles((theme) => ({
+  person: {
+    paddingTop: 15,
+    textAlign: 'center',
+    color: theme.palette.text.secondary,
+    fontSize: 14,
+    '&:hover': {
+      '& $icons': {
+        opacity: 1,
+        transition: 'opacity .3s ease-in-out',
+        MozTransition: 'opacity .3s ease-in-out',
+        WebkitTransition: 'opacity .3s ease-in-out',
+      },
+    },
+  },
+  portrait: {
+    position: 'relative',
+    margin: 0,
+    paddingBottom: 0,
+    width: '100%',
+    height: 'auto',
+    overflow: 'hidden',
+  },
+  name: {
+    fontWeight: 500,
+    paddingBottom: 2,
+  },
+  position: {
+    fontWeight: 400,
+    fontSize: 12,
+  },
+  icon: {
+    cursor: 'pointer',
+    transition: 'all .15s ease-in-out',
+    width: 30,
+    height: 30,
+    marginLeft: '0.5em',
+    marginRight: '0.5em',
+    '&:hover': {
+      transform: 'scale(1.15)',
+    },
+    borderRadius: 12,
+    backgroundColor: 'white',
+  },
+ 
+  icons: {
+    display: 'flex',
+    justifyContent: 'center',
+    alignItems: 'center',
+    paddingBottom: '0.5em'
+  },
+  cardText: {
+    padding: 5,
+    fontSize: 12,
+  },
+}));
+
+interface MemberProps {
+  readonly name: string;
+  position: string;
+  image: string;
+  linkedin: string;
+  email: string;
+}
+
+const Member: React.FC<MemberProps> = ({
+  name,
+  position,
+  image,
+  linkedin,
+  email,
+}) => {
+  const classes = useStyles();
+  return (
+    <Grid item md={4}>
+      <Paper className={classes.person} elevation={0}>
+
+        <LazyLoadImage className={classes.portrait} src={image} alt={name} />
+        <div className={classes.cardText}>
+          <Typography className={classes.name} variant="subtitle2">
+            {name}
+          </Typography>
+          <Typography className={classes.position} variant="subtitle2">
+            {position}
+          </Typography>
+        </div>
+        <div className={classes.icons}>
+        <a href={linkedin} rel="noopener noreferrer" target="_blank">
+            <img
+              className={classes.icon}
+              src="https://c4cneu-public.s3.us-east-2.amazonaws.com/Board/2020/icons/linkedin_icon.svg"
+              alt="Linkedin"
+            />
+          </a>
+          <a href={'mailto:' + email} rel="noopener noreferrer" target="_blank">
+            <img
+              className={classes.icon}
+              src="https://c4cneu-public.s3.us-east-2.amazonaws.com/Board/2020/icons/mail_icon.svg"
+              alt="Email"
+            />
+          </a>
+          </div>
+      </Paper>
+    </Grid>
+  );
+};
+
+export default Member;
