+++ src/pages/projects/Projects.tsx
@@ -1,10 +1,10 @@
 import React from 'react';
 import { Helmet } from 'react-helmet';
-import Hero from './Hero';
+import Hero from '../../components/Hero';
 import { Container, Grid } from '@material-ui/core';
 
 import ProjectCard from './ProjectCard';
-import { ReactComponent as ProjectsSvg } from '../svg/projects.svg';
+import { ReactComponent as ProjectsSvg } from '../../svg/projects.svg';
 
 const Projects: React.FC = () => {
   return (
